+++ Session-11/Library/TransactionLine.cs
@@ -13,14 +13,16 @@ namespace Library
         public DateTime Date { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public double Price { get; set; }
-        public double TotalCost { get; set; }        
-        public double TotalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal TotalCost { get; set; }        
+        public decimal TotalPrice { get; set; }
 
+        
         public TransactionLine()
         {
             ID = Guid.NewGuid();
             Date= DateTime.Now;
+            
         }
     }
 }
